refactor(mirror-registry): colocate command package
