+++ internal/mirrorregistry/cmd/cmd.go
