+++ internal/mirrorregistry/cmd/command.go
