+++ internal/mirrorregistry/cmd/command_test.go
