+++ internal/mirrorregistry/cmd/init.go
