+++ internal/mirrorregistry/cmd/init_test.go
