+++ internal/mirrorregistry/cmd/install.go
