+++ internal/mirrorregistry/cmd/migrate.go
