+++ internal/mirrorregistry/cmd/uninstall.go
