+++ internal/mirrorregistry/cmd/uninstall_test.go
